ai_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pillow==10.4.0
